Recreating API
- Removing dependency with UIView as base class for ViewBuilder, now your views will only contain the methods available. Before that, if you create a class like Spacer(), you could access UIKit methods like Spacer().backgroundColor = .black. But this cause a problem between the view builders and ui views, we're trying to solve that.

Because swift compiler doesn't handle very well with associatedtype, we create the basic protocol that is ViewCreator and the super with associatedtype View: UIView that is UIViewCreator. Using this way we have now more clear what is what. So, root views are made to extend only ViewCreator and finals views extend UIViewCreator. If you have a root view like SignupView, that methods available are only the ones that you write in your view class, you don't have any more a lot off unusable methods (this my change). Remember always to use the type cast method to have access to your UIView provided by as(_:) function.

- Preparing a new way to create your cells for table view and collection view.

- Renamed a bunch of classes and methods.

- Some classes or methods may be broken with this update