New CollectionEngine!
- Collection has now same constructor as tables, but the layout is separated. You can create your own layout by using the .layoutMaker function, it is not fully implemented and it could have some layout errors. CollectionFlow make everything works, but it is incapable of listing cells that are in the same group but in other scroll direction, like:

A, D, D
B, D, D
C, D, D

In this case A, B and C have 1/3 of width and 60 (example) of height, and D has 2/3 of width and 60*3 of height. The scroll direction is in vertical direction. The result calculated by UICollectionViewFlowLayout should be:

  , D, D
A, D , D
  , D, D
B, C,

It may be necessary to implement an especial CollectionViewLayout just to keep everything in the right order and rewrite the LayoutGroup and LayoutItem to calculate based on CGFrame not on CGSize, providing exact position for all cells. (Searching for solution)