ListManager
ListManager was refactored so it could be more precise. In the old versions, ListManager was only responsible for updating the section and rows but with no connection with the data source of the tableView, the UICList.Group was always recalculating the indexes and recreating the rows because of the generic class Element. Even if the old approach was some way to do this, ForEach was always calling the onAppear function causing a never end data reload.

Now, ListManager is responsible to update sections and rows and this is reflected in dataSource and, to keep the row actions working, it was added the ListManager.Delete and ListManager.Append states so table view can reload data without reflecting directly in the data source.

ForEach causes some bugs and to solve them ListManager forces the for each in the list or collection to reload before calling the reload data function.

After all changes, it was noticed a big reduction of memory and CPU usage.

About ListManager as a data source, it was changed how section, rows, header, and footer are. Now, we have SectionManager and RowManager. The first one has the rows, header and footer separated so it can be updated and found with no complexity. Footer and Header never update so even if it is a RowManager, it will always behave as non-dynamic content.  The second one has the content and update, if the content is dynamic, finding itself section and calling the ListManager to update the data.