AutoLayout for collection cells
Using the CollectionLayoutManager gives more control of all items inside the collection layout and this allows to implement more helpers that will help using collection layout. It was added the .estimated as constrain rule for layout and is updated every time the view inside the collection cell change sizes. When that happen, collection layout has to recalculate the cell sizes to be equal to layout manager.