Closes #4
UICResized will resize views given superview layout

Some views of UIKit can't use auto layout and this will cause a lot of constraints errors. So, we tested the UICResized in tableHeaderView, tableFooterView and backgroundView of the tableView. More tests will be done in the future versions.

If you are implementing your own views and need to use the UIViewWrapper to get the view of some view creator, you shouldn't use the UIKit addSubview function, because layout will breaks.

We made a lot of changes in these situations. The first one, you should always use the AddSubview struct and call addSubview to add views safety (this will not add constraints and will add subview by checking translate resizing mask to constraints). The second one, when you have a view that can't be resized with auto layout (like view of viewControllers, backgroundView of tableView and so many others), you should use the UICResized calling the onAdd(_:)  -> addSubview(_:), setting the height and width priority and finalizing with watch(in:) function.

In the second case, UICreator will watch superview, the view and the subview to properly set the size of the subview frame without breaking constraints.