Testing ReusableView with OperationQueue
Now ReusableView will not remove and add the same view that already is inside of it. Just for tests, it was added OperationQueue to verify if the populate process can be more faster than the old approach.

Adjusts collection to calculate size with content offset.

Added one more function to bind for Value class.