New Creator Engine
Layout:

Removed SnapKit as layout manager and added the EasyAnchor. By changing the creator engine caused some weird bugs on constraints and that only happened with constraints created with SnapKit. Using the UIKit's anchor, the problem stopped from happening and to use a better way to declare constraints, it was found the EasyAnchor library.

Render:

The render API manages the current state of any creator. In old versions, the commit on states happens when the superview commit on it watching views, causing a dependency from superview. In case of the superview is the root on layout hierarchy, there was no major superview to commit its states, causing a frozen behavior.

Render will keep the state updated and commit when it is convenient. In case the view needs to commit, Render will look for other subviews that needs too and so commit in everybody. But, after the view move to next state, if a new handler is added to the already committed state, render will put it in a queue and commit when UIKit execute the OperationQueue.

This worked really great and no views are being avoided. The only view that has to change a bit was the views created with View Representable protocol. Maybe there is a better solution for that, because creator, now, is adding a superview (Maker) to commits the handlers of view representable. So, if you need to hide the view, or remove from superview, always use the wrapper variable.

Tree:

Creator now has lazy load for view as an opportunity to improve the updates on creators on future versions. So, to keep the hierarchy that is used internally, the creators have a hierarchy of creators and you can print it using the .printCreatorTrace() in any UIView.

—

These updates comes with a research on declarative UI and so there are a lot of things that needs to change on UICreator.