UICreator Refactor
It was removed the dependency to TemplateView and merged the methods to a new protocol UICView. When the view is requested to load, UICreator will call body and instanciate the view inside the RootView as same as before.

-- Refactor on ReactiveCenter [Failing]

Value and Relay were not being used very well, Relay is only being used internally. With some big changes, all Value when is projected will be converted to Relay that provides a lot of methods like map(_:), bind(to:) and operators like <-> and <~>. With this change, the idea is to expand the functions on Relay to be more helpful.